refactor: moved functions for profile picture to separate file
